pyproject.toml
[project]
name = "drexel-sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "google-cloud-aiplatform>=1.130.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.1",
    "langgraph>=1.0.5",
    "pypdf>=6.5.0",
    "tavily-python>=0.7.17",
]
